requirements.txt
annotated-types==0.7.0
cffi==1.17.1
cramjam==2.9.0
cryptography==43.0.3
cx-Oracle==8.3.0
fastparquet==2024.11.0
fsspec==2024.10.0
greenlet==3.1.1
joblib==1.4.2
numpy==2.0.2
oracledb==2.5.0
packaging==24.2
pandas==2.2.3
polars==1.15.0
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.1
pydantic-settings==2.3.4
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
lightgbm==3.3.5
